Experimental support for doing W3C DC API over 18013-5 Session Encryption.
This is an **experimental** branch for using ISO/IEC 18013-5 transport
and session encryption to convey W3C Digital Credentials API requests
and responses.

In a nutshell, the mdoc may convey the ability to do this - including
which DC API it supports - using the new 18013-5 Second Edition
`Capabilities` mechanism. If the reader selects to use this mechanism
they simple include the `dcApiSelected` boolean in
`SessionEstablishment` and instead of exchanging `DeviceRequest` and
`DeviceResponse` CBOR, the W3C DC JSON request and response strings
are exchanged.

We also need to define the origin for this to work and this is set
simply to iso-18013-5://<sha256-of-encoded-session-transcript>.

This protocol works both with OpenID4VP 1.0 and - interestingly enough
- also ISO/IEC 18013-7 Annex C. See https://youtu.be/mvM6MpfljSw for
an example.

This is an experiment meant as a basis for discussion in various SDOs,
not an intent to merge any of this to the main branch.

Test: Manually tested.
Signed-off-by: David Zeuthen <zeuthen@google.com>